requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.2.0
pdf2image==1.16.3
reportlab==4.0.7
python-dotenv==1.0.0

